Added support for pretty URLs
Instead of generating a HTML file with the same name in the equivilant output path to the input, it will now create a directory with that name and an index.html file inside of that.

This means you can have /page as the URL instead of /page.html.

Still want /page.html style of URLs? Pass a third boolean parameter indicating your choice. True meaning /page and false meaning /page.html. If not provided, it will default to true and generate /page style URLs.